move inputs-related files to 'inputs' folder
