+++ src/inputs/keyboard-configs.ts
