+++ src/inputs/ui-inputs.ts
@@ -1,8 +1,8 @@
 import { globalScene } from "#app/global-scene";
-import type { InputsController } from "#app/inputs-controller";
 import { GAME_SPEEDS } from "#constants/app-constants";
 import { Button } from "#enums/button";
 import { UiMode } from "#enums/ui-mode";
+import type { InputsController } from "#inputs/inputs-controller";
 import { settings } from "#system/settings-manager";
 import { AudioSettingsUiHandler } from "#ui/audio-settings-ui-handler";
 import { DisplaySettingsUiHandler } from "#ui/display-settings-ui-handler";
