+++ tsconfig.json
@@ -37,7 +37,7 @@
       "#field/*": ["./field/*.ts"],
       "#form-change-triggers/*": ["./data/species-form-change-triggers/*.ts"],
       "#init/*": ["./data/init/*.ts"],
-      "#inputs/*": ["./configs/inputs/*.ts"],
+      "#inputs/*": ["./inputs/*.ts"],
       "#items/*": ["./data/items/*.ts"],
       "#modifier/*": ["./modifier/*.ts"],
       "#moves/*": ["./data/moves/move-attrs/*.ts", "./data/moves/move-conditions/*.ts", "./data/moves/*.ts"],
